backend/pyproject.toml
[project]
name = "insurancecopilotai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "langchain>=1.2.8",
    "langchain-core>=1.2.8",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph-checkpoint-postgres>=3.0.4",
    "openai>=2.16.0",
    "psycopg-binary>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.8.0",
    "bcrypt==4.0.1",
    "sqlalchemy>=2.0.46",
    "tenacity>=9.1.2",
    "uvicorn>=0.40.0",
    "flashrank>=0.2.10",
    "qdrant-client>=1.16.2",
    "fastembed>=0.7.4",
]
